Get one Wikimapia object by ID

Retrieves single Wikimapia object by ID.

Usage

wm_get_by_id_single(
  id,
  language = "en",
  wm_api_key = getOption("wikimapia_api_key", default = "example"),
  data_blocks = "main,geometry,edit,location,attached,photos,comments,translate"
)

Arguments

id

a numeric or characted vector of length 1 with Wikimapia object ID to fetch. You can get a list of objects to fetch using wm_get_from_bbox

language

Wikimapia language to retrieve. This is specified language in ISO 639-1 format. Default language is 'en'.

wm_api_key

Your wikimapia API Key. If not specified, the default 'example' key is used, however it is limited to about 1 request per 30 seconds and is for testing purposes only.

data_blocks

The comma-separated blocks of data you want to return: main, geometry, edit, location, attached, photos, comments, translate, similar_places, nearest_places, nearest_comments, nearest_streets, nearest_hotels. Default is "main,geometry,edit,location,attached,photos,comments,translate".

Value

`sf` object with nested details.

Details

Uses [`place.getbyid`](http://wikimapia.org/api#placegetbyid) Wikimapia API function to retrieve full object details.